requirements.txt
numpy
scikit-learn
torch
torchvision
torch-summary==1.4.5
matplotlib
pytest
PyYAML==6.0.1
click==8.1.7
pandas
seaborn
iisignature==0.24
xgboost
